# Fix(tokenizing): Use multi-core