This improves speed for tokenizer and adds a progress bar.

Benchmark:

| dataset | old | new | 
| --- | --- | --- |
| teknium/GPT4-LLM-Cleaned | ~70s | < 10s |
| WizardLM/WizardLM_evol_instruct_V2_196k | ~5 min | ~55s | 

Note: just internally counting, nothing strict.

Future PR for more speed can:
- use batching. (requires changing class to not use `iter`)
- fix Dataset.from_list and shuffle (do we need to save to a list? why not save to a dataset directly to cut off this part)
- ConstantIterList can be improved

**Full credit to**: neverendingtoast