@NanoCode012 2023-07-19
One note: this removes the error if dataset is empty. We can simply just add into it the `data.py` file instead.